+++ src/nsidc/icesat2gis/read_geom.py
@@ -1,3 +1,4 @@
+from collections.abc import Sequence
 from pathlib import Path
 from typing import Literal, cast, get_args
 
@@ -14,8 +15,24 @@ from nsidc.icesat2gis.exceptions import IceSatMissingDataError
 GroundTrack = Literal["gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"]
 
 
+# Default ground_track core variables for ATL08.
+ATL08_DEFAULT_GT_CORE_VARS = (
+    "canopy/h_canopy",
+    "canopy/h_canopy_uncertainty",
+    "canopy/h_median_canopy",
+    "canopy/photon_rate_can",
+    "terrain/h_te_best_fit",
+    "terrain/h_te_uncertainty",
+    "terrain/photon_rate_te",
+    "terrain/terrain_slope",
+)
+
+
 def _read_points_for_gt(
-    *, ground_track: GroundTrack, filepath: Path
+    *,
+    ground_track: GroundTrack,
+    filepath: Path,
+    variables_to_include: Sequence[str],
 ) -> gpd.GeoDataFrame:
     """Reads 100m segment points from ATL08 for the given ground track.
 
@@ -36,17 +53,25 @@ def _read_points_for_gt(
         print(msg)
         raise IceSatMissingDataError(msg) from e
 
+    # Extract variables
     lats = ds.latitude
     lons = ds.longitude
-    canopy_heights = ds.canopy.h_canopy
     delta_time = ds.delta_time
 
+    variables = {}
+    for var_path in variables_to_include:
+        var_name = var_path.rsplit("/", maxsplit=1)[-1]
+        variables[var_name] = ds[var_path]
+
+    # Construct gdf
     gdf = gpd.GeoDataFrame(
         data={
+            # Reference info
             "ground_track": [ground_track] * len(lons),
             "source_filename": [filepath.name] * len(lons),
-            "h_canopy": canopy_heights,
             "delta_time": delta_time,
+            # User-provided variables
+            **variables,
         },
         geometry=gpd.points_from_xy(lons, lats),
         crs="EPSG:4326",
@@ -59,12 +84,20 @@ def _read_points_for_gt(
     return gdf
 
 
-def read_points_from_atl08(*, filepath: Path) -> gpd.GeoDataFrame:
+def read_points_from_atl08(
+    *,
+    filepath: Path,
+    gt_variables_to_include: Sequence[str] = ATL08_DEFAULT_GT_CORE_VARS,
+) -> gpd.GeoDataFrame:
     """Return a GeoDataFrame containing points representing ground tracks."""
     gdfs = []
     for ground_track in get_args(GroundTrack):
         try:
-            gdf = _read_points_for_gt(ground_track=ground_track, filepath=filepath)
+            gdf = _read_points_for_gt(
+                ground_track=ground_track,
+                filepath=filepath,
+                variables_to_include=gt_variables_to_include,
+            )
             gdfs.append(gdf)
         except IceSatMissingDataError:
             continue
@@ -176,7 +209,7 @@ def lines_from_atl08_points(
     """
     geod = Geod(ellps="WGS84")
     multi_linestrings = {}
-    for ground_track in ("gt1l", "gt1r", "gt2l", "gt2r", "gt3l", "gt3r"):
+    for ground_track in set(points.ground_track):
         points_for_track = points[points.ground_track == ground_track].copy()
 
         # Distances between consecutive pairs in meters
@@ -219,10 +252,6 @@ def lines_from_atl08_points(
         # Track multilinestring and attrs per ground track
         multi_linestrings[ground_track] = {
             "geometry": multi_line,
-            "h_canopy_mean": points_for_track.h_canopy.mean(),
-            "h_canopy_min": points_for_track.h_canopy.min(),
-            "h_canopy_max": points_for_track.h_canopy.max(),
-            "h_canopy_std": points_for_track.h_canopy.std(),
             "delta_time_start": points_for_track.delta_time.min(),
             "delta_time_end": points_for_track.delta_time.max(),
         }
@@ -232,18 +261,6 @@ def lines_from_atl08_points(
             "ground_track": list(multi_linestrings.keys()),
             "source_filename": [list(set(points.source_filename))[0]]
             * len(multi_linestrings),
-            "h_canopy_min": [
-                line["h_canopy_min"] for line in multi_linestrings.values()
-            ],
-            "h_canopy_max": [
-                line["h_canopy_max"] for line in multi_linestrings.values()
-            ],
-            "h_canopy_mean": [
-                line["h_canopy_mean"] for line in multi_linestrings.values()
-            ],
-            "h_canopy_std": [
-                line["h_canopy_std"] for line in multi_linestrings.values()
-            ],
             "delta_time_start": [
                 line["delta_time_start"] for line in multi_linestrings.values()
             ],